Placing functions in correct files
