+++ model/trace_test.go
@@ -67,68 +67,3 @@ func TestTraceNormalizeTimestamps(t *testing.T) {
 	assert.Equal(t, span.StartTime, tt1.UTC())
 	assert.Equal(t, span.Logs[0].Timestamp, tt2.UTC())
 }
-
-func TestFlattenToSpanMaps(t *testing.T) {
-	b1 := &model.Trace{
-		Spans: []*model.Span{
-			{TraceID: model.NewTraceID(1, 2), SpanID: model.NewSpanID(1), OperationName: "x"},
-			{TraceID: model.NewTraceID(1, 3), SpanID: model.NewSpanID(2), OperationName: "y"},
-		},
-	}
-
-	b2 := &model.Trace{
-		Spans: []*model.Span{
-			{TraceID: model.NewTraceID(1, 2), SpanID: model.NewSpanID(2), OperationName: "z"},
-		},
-	}
-
-	t1 := []*model.Span{
-		{TraceID: model.NewTraceID(1, 2), SpanID: model.NewSpanID(1), OperationName: "x"},
-		{TraceID: model.NewTraceID(1, 2), SpanID: model.NewSpanID(2), OperationName: "z"}}
-
-	t2 := []*model.Span{{TraceID: model.NewTraceID(1, 3), SpanID: model.NewSpanID(2), OperationName: "y"}}
-	spanMap := make(map[model.TraceID][]*model.Span)
-	b1.FlattenToSpansMaps(spanMap)
-	b2.FlattenToSpansMaps(spanMap)
-	assert.Equal(t, t1, spanMap[model.NewTraceID(1, 2)])
-	assert.Equal(t, t2, spanMap[model.NewTraceID(1, 3)])
-}
-
-
-func TestDenormalizeProcess(t *testing.T){
-	t1 := &model.Trace{
-		Spans: []*model.Span{
-			{
-            TraceID: model.NewTraceID(1, 2), 
-            SpanID: model.NewSpanID(1), 
-            OperationName: "x",
-         },
-			{
-            TraceID: model.NewTraceID(1, 3), 
-            SpanID: model.NewSpanID(2), 
-            OperationName: "y",
-         },
-		},
-	}
-   p1 := model.NewProcess("process1", model.KeyValues{})
-
-	t2 := &model.Trace{
-		Spans: []*model.Span{
-			{
-            TraceID: model.NewTraceID(1, 2), 
-            SpanID: model.NewSpanID(1), 
-            OperationName: "x",
-            Process: model.NewProcess("process1", model.KeyValues{}),
-         },
-			{
-            TraceID: model.NewTraceID(1, 3), 
-            SpanID: model.NewSpanID(2), 
-            OperationName: "y",
-            Process: model.NewProcess("process1", model.KeyValues{}),
-         },
-		},
-	}
-   t1.DenormalizeProcess(p1)
-	assert.Equal(t, t1, t2)
-
-}
